+++ icu_pipeline/unit/quantity.py
@@ -5,10 +5,15 @@ from icu_pipeline.unit.converter import BaseConverter
 
 
 class UnitConverter(BaseConverter):
-    SI_UNIT = "unit" # Abstract Unit. Usually used for unspecific medications (eg. Vasopressine)
-    AVAILABLE_UNITS = ["unit", "milliunit"]
+    SI_UNIT = "unit"  # Abstract Unit. Usually used for unspecific medications (eg. Vasopressine)
+    AVAILABLE_UNITS = ["unit", "milliunit", "%"]
 
-    def _convertToSI(self, source_unit: str, data: Series[Quantity], dependencies: dict[str,DataFrame]):
+    def _convertToSI(
+        self,
+        source_unit: str,
+        data: Series[Quantity],
+        dependencies: dict[str, DataFrame],
+    ):
         convert: Callable[[float], float] = lambda v: v
         # Data can have any Unit and will be transformed to °C
         match source_unit:
@@ -20,15 +25,20 @@ class UnitConverter(BaseConverter):
             case "milliunit":
                 convert = lambda v: v / 10e3
 
+            case "%":
+                convert = lambda v: v / 100
+
             # Not Implemented
             case _:
                 raise NotImplementedError
 
-        return data.apply(lambda q: Quantity(
-            value=convert(q["value"]),
-            unit=self.SI_UNIT))
+        return data.apply(
+            lambda q: Quantity(value=convert(q["value"]), unit=self.SI_UNIT)
+        )
 
-    def _convertToTarget(self, sink_unit: str, data: Series[Quantity], dependencies: dict[str,DataFrame]):
+    def _convertToTarget(
+        self, sink_unit: str, data: Series[Quantity], dependencies: dict[str, DataFrame]
+    ):
         convert: Callable[[float], float] = lambda v: v
         # Data uses °C and can be transformed in to any Unit
         match sink_unit:
@@ -40,10 +50,11 @@ class UnitConverter(BaseConverter):
             case "milliunit":
                 convert = lambda v: v * 10e3
 
+            case "%":
+                convert = lambda v: v * 100
+
             # Not Implemented
             case _:
                 raise NotImplementedError
 
-        return data.apply(lambda q: Quantity(
-            value=convert(q["value"]),
-            unit=sink_unit))
+        return data.apply(lambda q: Quantity(value=convert(q["value"]), unit=sink_unit))
